+++ spec/lib/danger/commands/plugin_lint_spec.rb
@@ -1,4 +1,4 @@
-require "danger/commands/plugins/plugin_lint"
+require "danger/commands/plugins/plugin_json"
 
 module Danger
   describe Danger::PluginLint do
@@ -8,7 +8,7 @@ module Danger
 
     it "runs the command" do
       # allow(STDOUT).to receive(:puts) # this disables puts
-      Danger::PluginLint.run(["spec/fixtures/plugins/example_fully_documented.rb"])
+      Danger::PluginJSON.run(["spec/fixtures/plugins/example_fully_documented.rb"])
     end
   end
 end